requirements_training.txt
# Requirements for MedMNIST Training Pipeline
# ==========================================

# Core PyTorch and Lightning
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.8.0

# MedMNIST Dataset
medmnist>=2.1.0

# Data Processing and Visualization
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration Management
PyYAML>=6.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.13.0

# Optional: Additional visualization libraries
# plotly>=5.0.0
# ipywidgets>=7.6.0

# Development and Testing
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0

# Jupyter for interactive development (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0


